+++ src/AccessControl/Owned.py
@@ -13,8 +13,6 @@
 """Support for owned objects
 """
 
-from zope.deferredimport import deprecated
-
 # BBB
 from AccessControl.owner import EditUnowned
 from AccessControl.owner import EmergencyUserCannotOwn
@@ -22,12 +20,3 @@ from AccessControl.owner import UnownableOwner
 from AccessControl.owner import absattr
 from AccessControl.owner import ownableFilter
 from AccessControl.owner import ownerInfo
-
-
-deprecated(
-    "The Owned class has moved to OFS.owner. This compatibility "
-    "shim will be removed in AccessControl version 5. Please "
-    "depend on Zope2 and import from OFS.owner or use the "
-    "new minimal Owned class from AccessControl.owner.",
-    Owned='OFS.owner:Owned',
-)