+++ src/AccessControl/Role.py
@@ -13,8 +13,6 @@
 """Access control support
 """
 
-from zope.deferredimport import deprecated
-
 # BBB
 from AccessControl.rolemanager import DEFAULTMAXLISTUSERS
 from AccessControl.rolemanager import _isBeingUsedAsAMethod
@@ -26,13 +24,3 @@ from AccessControl.rolemanager import gather_permissions
 from AccessControl.rolemanager import instance_attrs
 from AccessControl.rolemanager import instance_dict
 from AccessControl.rolemanager import reqattr
-
-
-deprecated(
-    "RoleManager has moved to OFS.role. Please "
-    "depend on Zope and import from OFS.role or use the new minimal "
-    "RoleManager class from AccessControl.rolemanager. "
-    "This backward compatibility shim will be removed in AccessControl "
-    "version 5.",
-    RoleManager='OFS.role:RoleManager',
-)